rerender_component
