+++ crates/zoon/src/relations.rs
@@ -2,7 +2,7 @@ use crate::{Element, block_call_stack::{__Block, __BlockCallStack}};
 use std::collections::HashSet;
 use crate::runtime::{RELATIONS, CACHES, LVARS, CVARS};
 use crate::tracked_call_stack::__TrackedCallStack;
-use crate::component::__ComponentData;
+use crate::component::{__ComponentData, rerender_component};
 use crate::log;
 
 #[derive(Clone, Copy, Hash, PartialEq, Eq)]
@@ -17,12 +17,12 @@ pub struct __Relations(HashSet<Relation>);
 impl __Relations {
     pub fn add_dependency(dependency: __Block) {
         if let Some(last_block) = __BlockCallStack::last() {
-            if let __Block::LVar(l_var_id) = &dependency {
-                log!("A add_dependency LVar");
-                if let __Block::Cmp(cmp_id) = &last_block {
-                    log!("B add_dependency LVar({:#?}) to CMP({:#?})", l_var_id, cmp_id);
-                }
-            }
+            // if let __Block::LVar(l_var_id) = &dependency {
+            //     log!("A add_dependency LVar");
+            //     if let __Block::Cmp(cmp_id) = &last_block {
+            //         log!("B add_dependency LVar({:#?}) to CMP({:#?})", l_var_id, cmp_id);
+            //     }
+            // }
 
             match last_block {
                 __Block::Cache(_) | __Block::Cmp(_)=> {
@@ -48,12 +48,12 @@ impl __Relations {
                 .do_get_dependents(block)
         });
 
-        if let __Block::LVar(l_var_id) = &block {
-            if dependents.len() > 0 {
-                log!("refresh LVar dependents {}, LVar({:#?})", dependents.len(), l_var_id);
-            }
-            // return;
-        }
+        // if let __Block::LVar(l_var_id) = &block {
+        //     if dependents.len() > 0 {
+        //         log!("refresh LVar dependents {}, LVar({:#?})", dependents.len(), l_var_id);
+        //     }
+        //     // return;
+        // }
 
         for block in dependents {
             match block {
@@ -74,30 +74,8 @@ impl __Relations {
                     }
                 }
                 __Block::Cmp(track_call_id) => {
-                    log("refresh CMP!");
-                     
-                    let component_creator = CVARS.with(|c_vars| {
-                        c_vars
-                            .borrow()
-                            .data::<__ComponentData>(&track_call_id)
-                            .clone()
-                    });
-                    if let Some(rcx) = component_creator.rcx {
-                        let parent_call_from_macro = component_creator.parent_call_from_macro.unwrap();
-                        parent_call_from_macro.borrow_mut().selected_index = component_creator.parent_selected_index_from_macro.unwrap();
-
-                        __TrackedCallStack::push(parent_call_from_macro);
-                        let mut cmp = (component_creator.creator)();
-                        __TrackedCallStack::pop();
-
-                        let parent_call = component_creator.parent_call.unwrap();
-                        parent_call.borrow_mut().selected_index = component_creator.parent_selected_index.unwrap() - 1;
-                       
-                        __TrackedCallStack::push(parent_call);
-                        cmp.render(rcx);
-                        __TrackedCallStack::pop();
-                        // rerender();
-                    }
+                    // log("refresh CMP!");
+                    rerender_component(track_call_id)
                 }
                 __Block::SVar(_) | __Block::LVar(_) => ()
             }
